chore: move StatusIndicator into icons lib to use with ChainIcon
