+++ libs/icons/src/StatusIndicator/StatusIndicator.tsx
