+++ libs/icons/src/StatusIndicator/index.ts
